+++ .gitignore
@@ -14,6 +14,9 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
+# zapggo binary
+/bin
+zapgo
 #zapgo test files
 zap.log
 zap-report*
