+++ cmd/clean.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"github.com/1azunna/zapgo/internal/zapgo"
